requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
SQLAlchemy==1.4.41
psycopg2-binary==2.9.5
pydantic==1.10.2
alembic==1.8.1
python-dotenv==0.20.0
pytest==7.1.2
asyncpg==0.26.0
